Moved Map to its own folder to try to fix conflict.
